requirements.txt
snowflake-snowpark-python[pandas]
snowflake-cli-labs==0.1.8
